+++ CHANGELOG.md
@@ -586,6 +586,7 @@
 * Added /iclock_bypass, allowing GM+ to let non-GMs in an area whose IC chat is locked to talk in IC. The effect disappears as soon as the target moves area or their area has their IC chat unlocked
 * Improved output of /blind, /deafen, /gag if no arguments are passed
 * Made /blind, /deafen, /gag, /bloodtrail echo the ID of the affected target as part of output message
+* GMs are no longer subject to the server music flood guard
 * Fixed scream_range in area list yaml files not supporting the keyword <ALL> to indicate all areas should be able to receive a scream coming from a particular area
 * Fixed scream_range in area list yaml files not checking if the areas a scream can reach to from a particular area exist
 * Fixed /scream, /whisper and /party_whisper raising errors if a message was sent to a deafened player with a bypass message starter. They now sent messages but filtered
@@ -617,5 +618,6 @@
 * Fixed /charselect notifying of the wrong person running the command to officers
 * Fixed output of /scream_range being formatted different from /minimap. It now lists areas in order by ID with the format number-name
 * Fixed /scream, /whisper and /party_whisper not sending character folder information, which prevented rendering of showname images
+* Fixed /play bypassing IC mutes, blockdj and the server music flood guard
 * Removed deprecated AO commands, and deprecated packets opKICK and opBAN
 * Dropped Python 3.6 support